Some updates. Also I'm just including the required DLLs for making the Roguelike example work off the bat, even though in the future we should probably require that people download the third-parties themselves...

+++ trunk/depends/Standard/Examples/Roguelike/include/color.hpp
@@ -145,7 +145,7 @@ public :
 	@LuaEx myColor = tcod.Color(24,24,255)
 	*/
 	TCODColor(uint8 r, uint8 g, uint8 b): r(r), g(g), b(b) {}
-	TCODColor(int r, int g, int b): r(r), g(g), b(b) {}
+	TCODColor(int r, int g, int b): r((uint8)r), g((uint8)g), b((uint8)b) {}
 	TCODColor(const TCOD_color_t &col): r(col.r), g(col.g), b(col.b) {}
 	TCODColor(float h, float s, float v);
